src/travel_planner_mcp/requirements.txt
openai>=1.0.0
python-dotenv
pydantic>=2.0
fastapi
uvicorn
enrichmcp==0.4.7
mcp-use==1.3.7
